fix: Copilotレビュー指摘に基づきRoom/非同期処理を改善

+++ android/app/src/main/java/com/anagram/analyzer/data/db/AnagramDao.kt
@@ -14,5 +14,5 @@ interface AnagramDao {
     suspend fun lookupWords(sortedKey: String): List<String>
 
     @Query("SELECT COUNT(*) FROM anagram_entries")
-    suspend fun count(): Int
+    suspend fun count(): Long
 }
